Finished adding support for the change log on the issueForm.jsp. Added admin checking to IssueForm.buildSQLQuery to make it faster for admins. Added support for metrics coming from EOI. Made batching optional in DefaultDataLoader. Fixed missing table errors when attempting to drop tables that don't exist. Added an experimental 'last request took xx ms' statistic that is displayed in the footer.

+++ src/main/java/net/ehicks/bts/Startup.java
@@ -65,7 +65,7 @@ public class Startup
         long subTaskStart = System.currentTimeMillis();
         int tablesCreated = 0;
         for (DBMap dbMap : DBMap.dbMaps)
-            if (!EOI.isTableExists(dbMap))
+            if (!EOI.isTableExists(dbMap.tableName))
             {
                 String createTableStatement = SQLGenerator.getCreateTableStatement(dbMap);
                 EOI.executeUpdate(createTableStatement);
@@ -85,7 +85,11 @@ public class Startup
         {
             try
             {
-                EOI.executeUpdate("drop table " + dbMap.tableName);
+                if (EOI.isTableExists(dbMap.tableName))
+                {
+                    log.debug("Dropping " + dbMap.tableName + "...");
+                    EOI.executeUpdate("drop table " + dbMap.tableName);
+                }
             }
             catch (Exception e)
             {
